chore: move errorhandler to middleware
